+++ assess_acceptability_judgments/constituency_parser.py
@@ -1,8 +1,7 @@
+import json
 import os
-import zipfile
-from pathlib import Path
+import pkgutil
 from typing import List, Union, Optional
-from urllib.request import urlretrieve
 
 import benepar
 import nltk
@@ -14,8 +13,7 @@ from stanza.models.constituency.tree_reader import read_trees
 from supar import Parser
 from tqdm import tqdm
 
-from .ressources import CACHE_PATH, CORENLP_URL
-from .util import DownloadProgressBar
+from .core_nlp_parser_interface import CoreNLPParser
 
 
 # Add encoding of the parse tree using Tree-LSTM:
@@ -23,54 +21,36 @@ from .util import DownloadProgressBar
 # similar to https://www.hindawi.com/journals/cin/2022/4096383/
 
 
-class ConstituencyParserCoreNLP:
-    # Path to the corenlp JAR models to use for parsing and create Tree
-    # As of july 2023, Stanza does not return a Tree by a dictionary. Thus, we use NLTK API
-    # that parse and return a dependency parse tree.
-    CORENLP_DIRECTORY = "stanford-corenlp-full-2018-02-27"
-    JAR_FILE_NAME = os.path.join(CORENLP_DIRECTORY, "stanford-corenlp-3.9.1.jar")
-    JAR_MODEL_FILE_NAME = os.path.join(CORENLP_DIRECTORY, "stanford-corenlp-3.9.1-models.jar")
-
+class ConstituencyParserCoreNLP(CoreNLPParser):
     def __init__(self, verbose: bool = True, cache_path: Optional[str] = None) -> None:
         """
          Create a constituency parsing model that use CoreNLP constituency parser. To do so, we download the latest
         model from CoreNLP (i.e. 2018) as suggest by this Wiki
         https://github.com/nltk/nltk/wiki/Stanford-CoreNLP-API-in-NLTK.
 
+        The tag ar those suggested by this source https://gist.github.com/nlothian/9240750.
+
         :param verbose: (bool) Either or not to be verbose during the download of CoreNLP model. Default to `True`.
         :param cache_path: (Optional[str]) Optional parameter to set a cache path to download the CoreNP model to.
             If the cache_path is not set, the model are downloaded in the default cache path i.e. `'.cache/aaj'`.
         """
+        super().__init__(verbose, cache_path)
 
-        if cache_path is None:
-            cache_path = CACHE_PATH
-
-        self.jar_file_name = os.path.join(cache_path, self.JAR_FILE_NAME)
-        self.jar_model_file_name = os.path.join(cache_path, self.JAR_MODEL_FILE_NAME)
-
-        self.verbose = verbose
-        if not os.path.exists(self.jar_file_name) and not os.path.exists(self.jar_model_file_name):
-            if self.verbose:
-                reporthook = DownloadProgressBar()
-            else:
-                reporthook = None
-
-            # Download zipped file with verbose report
-            local_filename, _ = urlretrieve(CORENLP_URL, reporthook=reporthook)
-
-            # Create .cache directory if it does not exist
-            Path(cache_path).mkdir(parents=True, exist_ok=True)
-
-            # Unzip the file into the cache directory
-            with zipfile.ZipFile(local_filename, "r") as f:
-                f.extractall(cache_path)
+        data = pkgutil.get_data(
+            __name__,
+            os.path.join(
+                "./resources",
+                "treebank_tags_mapping.json",
+            ),
+        )
+        self._tags_mapping = json.loads(data.decode("utf-8"))
 
     def tree_parser_sentences(self, sentences: List[str]) -> List[List[Union[str, nltk.tree.tree.Tree]]]:
         """
         Method to parse sentences into constituency tree.
 
         :param sentences: (list) A list of sentence to parse into trees.
-        :return: A list of Stanza parse tree.
+        :return: A list of NLTK parse tree.
         """
         with CoreNLPServer(path_to_jar=self.jar_file_name, path_to_models_jar=self.jar_model_file_name) as server:
             parser = CoreNLPParser(url=server.url)